String Class: String similarity metrix comparision to defined threshold required Pandas Series
Comparison of the similarity metric to the defined threshold takes advantage of
Pandas.Series.where method. However for the cosine and qgram algorithms,
the return similarity metrics are a numpy array.
This causes an error.

AttributeError: 'numpy.ndarray' object has no attribute 'where'

Adding a check and conversion to ensure similarity metrics are
pandas Series when comparing to the threshold.